+++ src/main/java/dandi/dandi/weatherbatch/application/runner/WeatherBatch.java
@@ -1,6 +1,5 @@
 package dandi.dandi.weatherbatch.application.runner;
 
-import dandi.dandi.batchcommons.exception.BatchException;
 import dandi.dandi.batchcommons.jobparameter.DateTimeJobParameter;
 import dandi.dandi.weather.application.port.out.WeatherPersistencePort;
 import dandi.dandi.weather.application.port.out.WeatherRequestFatalException;
@@ -8,7 +7,6 @@ import dandi.dandi.weather.application.port.out.WeatherRequestRetryableException
 import dandi.dandi.weather.application.port.out.WeatherRequester;
 import dandi.dandi.weather.domain.WeatherLocation;
 import dandi.dandi.weather.domain.Weathers;
-import org.jetbrains.annotations.NotNull;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
@@ -18,20 +16,15 @@ import org.springframework.batch.core.configuration.annotation.StepScope;
 import org.springframework.batch.item.ItemProcessor;
 import org.springframework.batch.item.ItemReader;
 import org.springframework.batch.item.ItemWriter;
-import org.springframework.batch.item.database.PagingQueryProvider;
-import org.springframework.batch.item.database.builder.JdbcPagingItemReaderBuilder;
-import org.springframework.batch.item.database.support.SqlPagingQueryProviderFactoryBean;
 import org.springframework.batch.item.support.builder.CompositeItemWriterBuilder;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.retry.RetryPolicy;
 import org.springframework.retry.backoff.FixedBackOffPolicy;
 import org.springframework.retry.policy.SimpleRetryPolicy;
 
-import javax.sql.DataSource;
 import java.io.IOException;
 import java.net.SocketException;
 import java.util.List;
@@ -41,25 +34,22 @@ import java.util.stream.Collectors;
 @Configuration
 public class WeatherBatch {
 
-    private static final String JOB_NAME = "weatherBatchJob";
+    public static final String WEATHER_BATCH_JOB_NAME = "weatherBatchJob";
 
     private final JobBuilderFactory jobBuilderFactory;
     private final StepBuilderFactory stepBuilderFactory;
-    private final DataSource dataSource;
     private final WeatherRequester weatherRequester;
     private final DateTimeJobParameter dateTimeJobParameter;
     private final WeatherPersistencePort weatherPersistencePort;
 
     public WeatherBatch(JobBuilderFactory jobBuilderFactory, StepBuilderFactory stepBuilderFactory,
-                        DataSource dataSource, WeatherRequester weatherRequester,
-                        @Qualifier(value = "weatherBatchJobBaseDateTimeParameter") DateTimeJobParameter dateTimeJobParameter,
-                        WeatherPersistencePort weatherPersistencePort) {
+                        WeatherRequester weatherRequester, WeatherPersistencePort weatherPersistencePort,
+                        @Qualifier(value = "weatherBatchJobBaseDateTimeParameter") DateTimeJobParameter dateTimeJobParameter) {
         this.jobBuilderFactory = jobBuilderFactory;
         this.stepBuilderFactory = stepBuilderFactory;
-        this.dataSource = dataSource;
         this.weatherRequester = weatherRequester;
-        this.dateTimeJobParameter = dateTimeJobParameter;
         this.weatherPersistencePort = weatherPersistencePort;
+        this.dateTimeJobParameter = dateTimeJobParameter;
     }
 
     @Bean
@@ -69,23 +59,23 @@ public class WeatherBatch {
         return new DateTimeJobParameter(baseDateTime);
     }
 
-    @Bean(JOB_NAME)
+    @Bean(WEATHER_BATCH_JOB_NAME)
     public Job weatherBatch() throws IOException {
-        return jobBuilderFactory.get(JOB_NAME)
-                .start(weatherBatchStep(null, null))
+        return jobBuilderFactory.get(WEATHER_BATCH_JOB_NAME)
+                .start(weatherBatchStep(null, null, null))
                 .build();
     }
 
     @Bean
     @JobScope
     public Step weatherBatchStep(@Value("#{jobParameters[backOffPeriod]}") Long backOffPeriod,
-                                 @Value("#{jobParameters[chunkSize]}") Long chunkSize) {
+                                 @Value("#{jobParameters[chunkSize]}") Long chunkSize,
+                                 ItemReader<WeatherLocation> weatherLocationItemReader) {
         FixedBackOffPolicy fixedBackOffPolicy = new FixedBackOffPolicy();
         fixedBackOffPolicy.setBackOffPeriod(backOffPeriod);
         return stepBuilderFactory.get("weatherInsertion")
                 .<WeatherLocation, Weathers>chunk(chunkSize.intValue())
-                .reader(weatherLocationItemReader(null))
-                .processor(weatherItemProcessor())
+                .reader(weatherLocationItemReader)
                 .writer(itemWriters())
                 .faultTolerant()
                 .retryPolicy(retryPolicy())
@@ -99,40 +89,6 @@ public class WeatherBatch {
                 SocketException.class, true,
                 WeatherRequestFatalException.class, false);
         return new SimpleRetryPolicy(3, retryableExceptions);
-    }
-
-    @Bean
-    @StepScope
-    public ItemReader<WeatherLocation> weatherLocationItemReader(@Value("#{jobParameters[chunkSize]}") Long chunkSize) {
-        return new JdbcPagingItemReaderBuilder<WeatherLocation>()
-                .name("unusedImageItemReader")
-                .dataSource(dataSource)
-                .pageSize(chunkSize.intValue())
-                .fetchSize(chunkSize.intValue())
-                .queryProvider(pagingQueryProvider())
-                .rowMapper(rowMapper())
-                .build();
-    }
-
-    private PagingQueryProvider pagingQueryProvider() {
-        SqlPagingQueryProviderFactoryBean queryProvider = new SqlPagingQueryProviderFactoryBean();
-        queryProvider.setDataSource(dataSource);
-        queryProvider.setSelectClause("weather_location_id, x, y");
-        queryProvider.setFromClause("from weather_location");
-        queryProvider.setSortKey("weather_location_id");
-        try {
-            return queryProvider.getObject();
-        } catch (Exception e) {
-            throw BatchException.queryProviderObjectCreationFailed(JOB_NAME);
-        }
-    }
-
-    private RowMapper<WeatherLocation> rowMapper() {
-        return (rs, rowNum) -> new WeatherLocation(
-                rs.getLong("weather_location_id"),
-                rs.getInt("x"),
-                rs.getInt("y")
-        );
     }
 
     @Bean
